+++ src/sparsezoo/__init__.py
@@ -1,3 +1,3 @@
 from .api import *
-from .schemas import *
+from .models import *
 from .utils import *
