+++ src/sparsezoo/models/sparse_zoo_object.py
