+++ src/sparsezoo/models/tag.py
@@ -2,7 +2,7 @@
 Code related to model repo tags
 """
 
-from sparsezoo.schemas.object_schema import SparseZooObject
+from sparsezoo.models.sparse_zoo_object import SparseZooObject
 
 
 __all__ = ["Tag"]
