--- src/sparsezoo/schemas/__init__.py
@@ -1,8 +1,0 @@
-from .object_schema import *
-from .file_schema import *
-from .tag_schema import *
-from .optimization_schema import *
-from .release_version_schema import *
-from .result_schema import *
-from .user_schema import *
-from .model_schema import *
